some cleanup

+++ IEnumerableExtenders/Dto/FilteredDataResult.cs
@@ -3,7 +3,6 @@
 public class FilteredDataResult<T>
 {
     public List<T> Data { get; set; } = null!;
-    public int TotalCount { get; set; }
-    
+    public long TotalCount { get; set; }
     public Dictionary<string, object?> Aggregates { get; set; } = null!;
 }
